requirements.txt

aiohttp
discord.py
python-dotenv
Pillow
google-cloud-aiplatform
